pyproject.toml
[tool.poetry]
name = "Mumee"
version = "0.0.6"
description = "MUsic MEtadata Explorer"
authors = ["Billuc <billuc@hotmail.fr>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mumee"}]

keywords = ["mumee", "metadata", "python", "song metadata", "spotify", "youtube", "youtube music", "music", "song"]

homepage = "https://github.com/Billuc/Mumee"
repository = "https://github.com/Billuc/Mumee"
documentation = "https://github.com/Billuc/Mumee"

classifiers = [
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Utilities",
]

include = ["mumee/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
spotipy = "^2.23.0"
rapidfuzz = "^3.1.1"
python-slugify = "^8.0.1"
ytmusicapi = "^1.0.2"
taipan-di = "^0.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
vcrpy = "^4.3.1"
black = "^23.3.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "mumee"
]

[tool.black]
line-length = 92
